userspackage/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.1.0
Django==4.2.1
django-bootstrap-v5==1.0.11
django-environ==0.10.0
django-jazzmin==2.6.0
django-on-heroku==1.1.2
docutils==0.20.1
fontawesomefree==6.4.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
keyring==24.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.14
packaging==23.1
Pillow==9.5.0
pip-autoremove==0.10.0
pkginfo==1.9.6
psycopg2-binary==2.9.7
Pygments==2.16.1
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.3
soupsieve==2.4.1
sqlparse==0.4.4
twine==4.0.2
typing_extensions==4.7.0
urllib3==2.0.5
whitenoise==6.4.0
zipp==3.17.0
